agservice/authentication.py
# myapp/authentication.py
from datetime import datetime, timedelta
from django.utils import timezone
from rest_framework.authentication import TokenAuthentication
from rest_framework.exceptions import AuthenticationFailed

class ExpiringTokenAuthentication(TokenAuthentication):
    def authenticate_credentials(self, key):
        model = self.get_model()
        try:
            token = model.objects.select_related('user').get(key=key)
        except model.DoesNotExist:
            raise AuthenticationFailed('Invalid token')

        if not token.user.is_active:
            raise AuthenticationFailed('User inactive or deleted')

        # Check if the token has expired
        if token.created < timezone.now() - timedelta(seconds=10):
            raise AuthenticationFailed('Token has expired')

        return (token.user, token)
