requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
django-background-task==0.1.8
django-cron==0.6.0
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
icalendar==5.0.7
idna==3.4
install==1.3.5
kombu==5.3.0
logging==0.4.9.6
multidict==6.0.4
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
schedule==1.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
